Fixed ColorImage binary resize problem
see issue #22

+++ perception/image.py
@@ -1129,8 +1129,8 @@ class ColorImage(Image):
         :obj:`ColorImage`
             The resized image.
         """
-        resized_data = imresize(self.data, size, interp=interp).astype(np.uint8)
-        return ColorImage(resized_data, self._frame)
+        resized_data = imresize(self.data.astype(np.float32), size, interp=interp)
+        return ColorImage(resized_data.astype(np.uint8), self._frame)
 
     def find_chessboard(self, sx=6, sy=9):
         """Finds the corners of an sx X sy chessboard in the image.
